AidTypeVocabulary

The AidTypeVocabulary codelist defines a range of external codelists which themselves provide codes and descriptions for aid type. This is a Non-Core codelist.

AidTypeVocabulary

Format

A data frame with 4 rows and 6 variables:

code

double COLUMN_DESCRIPTION

name

character COLUMN_DESCRIPTION

description

character COLUMN_DESCRIPTION

category

logical COLUMN_DESCRIPTION

url

character COLUMN_DESCRIPTION

status

character COLUMN_DESCRIPTION

Source

https://iatistandard.org/en/iati-standard/203/codelists/

Examples

{
head(AidTypeVocabulary,10)
}

#> # A tibble: 4 x 6
#>    code name       description               category url                 status
#>   <dbl> <chr>      <chr>                     <lgl>    <chr>               <chr> 
#> 1     1 OECD DAC   List of codes provided b… NA       http://reference.i… active
#> 2     2 Earmarkin… This vocabulary has been… NA       http://reference.i… active
#> 3     3 Earmarkin… Codes A to L replicated … NA       https://reliefweb.… active
#> 4     4 Cash and … This vocabulary has been… NA       http://reference.i… active